Social: change channel id to int64

+++ go/src/socialapi/tests/integration.go
@@ -8,7 +8,7 @@ import (
 var (
 	ENDPOINT   = "http://localhost:8000"
 	ACCOUNT_ID = int64(1)
-	CHANNEL_ID = 1
+	CHANNEL_ID = int64(1)
 )
 
 func main() {
